@@ -2,7 +2,7 @@ package br.com.agenciaapp.agendamento.controller;
 
 import java.util.Optional;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
