@@ -1,7 +1,7 @@
 package br.com.agenciaapp.cadastroclinica.controller;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -40,7 +40,7 @@ public class AgendaController {
     @PostMapping("/{id}/profissional")
     public ResponseEntity<AgendaDto> cadastrar(
             @RequestBody @Valid AgendaDto dto,
-            @PathVariable(name = "id") Long id) {
+            @PathVariable Long id) {
         logger.info("cadastrar AgendaController");
         AgendaDto agendaDto = agendaService.criarAgenda(dto, id);
 
@@ -54,7 +54,7 @@ public class AgendaController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<AgendaDto> obterAgendaId(@PathVariable(name = "id") Long id) {
+    public ResponseEntity<AgendaDto> obterAgendaId(@PathVariable Long id) {
         logger.info("obterAgendaId AgendaController");
         AgendaDto agendaDto = agendaService.obterById(id);
 
