@@ -2,8 +2,8 @@ package br.com.agenciaapp.cadastroclinica.controller;
 
 import java.net.URI;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
